html/canvas/element/wide-gamut-canvas/2d.color.space.p3.and.uint8.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.color.space.p3.and.uint8</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.color.space.p3.and.uint8</h1>
<p class="desc">test getImageData with different seetting from display-p3 canvas</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("test getImageData with different seetting from display-p3 canvas");
_addTest(function(canvas, ctx) {

var color_style = 'rgb(50, 100, 150)';
var pixel_expected = [62, 99, 146, 255];
var epsilon = 2;
ctx.fillStyle = color_style;
ctx.fillRect(0, 0, 10, 10);

var pixel = ctx.getImageData(5, 5, 1, 1, {colorSpace: "display-p3", storageFormat: "uint8"}).data;
var passed = true;
_assertSame(pixel.length, pixel_expected.length, "pixel.length", "pixel_expected.length");
for (var i = 0; i < pixel.length; i++)
{
  if (Math.abs(pixel[i] - pixel_expected[i]) > epsilon) {
    passed = false;
  }
}
_assertSame(passed, true, "passed", "true");

pixel_expected = [50, 100, 150, 255];
var pixel = ctx.getImageData(5, 5, 1, 1, {colorSpace: "srgb", storageFormat: "uint8"}).data;
_assertSame(pixel.length, pixel_expected.length, "pixel.length", "pixel_expected.length");
for (var i = 0; i < pixel.length; i++)
{
  if (Math.abs(pixel[i] - pixel_expected[i]) > epsilon) {
    passed = false;
  }
}
_assertSame(passed, true, "passed", "true");


}, {colorSpace: "display-p3", pixelFormat: "uint8"});
</script>
